Move Apply components to new directory for Request Sandbox Access under consumer onboarding docs
