+++ src/containers/consumerOnboarding/components/sandbox/OAuthAppInfo.scss
