+++ src/containers/consumerOnboarding/components/sandbox/SelectedApis.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../../styles/variables.scss';
 @import '~@department-of-veterans-affairs/formation/sass/full';
 
 .apply-api-select {
