+++ src/containers/consumerOnboarding/components/sandbox/index.ts
@@ -1,0 +1,2 @@
+export * from './SandboxAccessSuccess';
+export * from './SandboxAccessForm';
